+++ test/lib/gas/GasTestBase.t.sol
@@ -37,12 +37,11 @@ abstract contract GasTestBaseStored {
   function getStored() internal view virtual returns (IMangrove, TestTaker, OLKey memory, uint);
 
   function printDescription() public virtual {
-    console.log("Description: %s", description);
+    printDescription("");
   }
 
   function printDescription(string memory postfix) public virtual {
-    description = string.concat(description, postfix);
-    printDescription();
+    console.log("Description: %s", string.concat(description, postfix));
   }
 
   function newOfferOnAllTestRatios() public virtual {